refact: productDetail 컴포넌트 분류, API 필요없는 부분 제거
