+++ src/Api/getProductId.js
@@ -2,7 +2,7 @@ import { instance } from "./Axios";
 
 export const getProductById = async (productId) => {
   try {
-    const response = await instance.get(`/products/${productId}`, {});
+    const response = await instance.get(`/products/${productId}`);
     return response.data;
   } catch (error) {
     console.error(error);
