+++ src/Api/getProducts.js
@@ -2,7 +2,7 @@ import { instance } from "./Axios";
 
 export const getProducts = async () => {
   try {
-    const response = await instance.get("/products", {});
+    const response = await instance.get("/products");
     return response.data;
   } catch (error) {
     console.error(error);
