+++ src/Api/loginUser.js
@@ -2,11 +2,7 @@ import { instance } from "./Axios";
 
 export const LoginUser = async (data) => {
   try {
-    const response = await instance.post(
-      "/auth/signIn",
-      JSON.stringify(data),
-      {}
-    );
+    const response = await instance.post("/auth/signIn", JSON.stringify(data));
     const { accessToken } = response.data;
     localStorage.setItem("accessToken", accessToken);
     window.location.href = "/items";
