+++ src/pages/AddItemPage.js
@@ -1,5 +1,3 @@
-import { useState } from "react";
-import styled from "styled-components";
 import AddItem from "../components/AddItem";
 
 const AddItemPage = () => {
